[#124] Refactor: BaseEntity 및 ErrorCode 이동
