+++ src/main/java/app/tamingo/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package app.tamingo;
+package app.tamingo.common.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.persistence.EntityListeners;
