+++ src/main/java/app/tamingo/domain/calendar/entity/ExternalTaskMapping.java
@@ -1,6 +1,6 @@
 package app.tamingo.domain.calendar.entity;
 
-import app.tamingo.BaseEntity;
+import app.tamingo.common.entity.BaseEntity;
 import app.tamingo.domain.calendar.enums.LinkStatus;
 import app.tamingo.domain.schedule.entity.Schedule;
 import jakarta.persistence.*;
