+++ src/main/java/app/tamingo/domain/gpt/service/common/GptService.java
@@ -1,24 +1,17 @@
 package app.tamingo.domain.gpt.service.common;
 
 import app.tamingo.common.exception.CustomException;
-import app.tamingo.common.response.gpt.GptErrorCode;
-import app.tamingo.common.webclient.ApiClient;
+import app.tamingo.domain.gpt.exception.GptErrorCode;
 import app.tamingo.domain.gpt.client.GptClient;
 import app.tamingo.domain.gpt.dto.ExampleGptResponse;
 import app.tamingo.domain.gpt.dto.GptRequest;
-import app.tamingo.domain.gpt.dto.GptResponse;
-import app.tamingo.domain.gpt.prompt.common.DataPrompt;
-import app.tamingo.domain.gpt.prompt.common.ExamplePrompt;
 import app.tamingo.domain.gpt.prompt.common.GeneralSystemPrompt;
 import app.tamingo.domain.gpt.prompt.common.PromptTemplate;
 import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.swagger.v3.oas.models.OpenAPI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.reactive.function.client.WebClient;
 
 @Service
 @Transactional
