+++ src/main/java/app/tamingo/domain/gpt/service/home/SuggestionGptService.java
@@ -1,7 +1,7 @@
 package app.tamingo.domain.gpt.service.home;
 
 import app.tamingo.common.exception.CustomException;
-import app.tamingo.common.response.gpt.GptErrorCode;
+import app.tamingo.domain.gpt.exception.GptErrorCode;
 import app.tamingo.domain.gpt.dto.GptRequest;
 import app.tamingo.domain.gpt.prompt.common.GeneralSystemPrompt;
 import app.tamingo.domain.gpt.prompt.common.PromptTemplate;
