+++ src/main/java/app/tamingo/domain/notificationsetting/entity/NotificationSetting.java
@@ -1,7 +1,7 @@
 package app.tamingo.domain.notificationsetting.entity;
 
 
-import app.tamingo.BaseEntity;
+import app.tamingo.common.entity.BaseEntity;
 import app.tamingo.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
