+++ src/main/java/app/tamingo/domain/schedule/entity/ScheduleAiLog.java
@@ -1,6 +1,6 @@
 package app.tamingo.domain.schedule.entity;
 
-import app.tamingo.BaseEntity;
+import app.tamingo.common.entity.BaseEntity;
 import app.tamingo.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
