+++ src/main/java/app/tamingo/domain/todo/entity/Todo.java
@@ -1,8 +1,7 @@
 package app.tamingo.domain.todo.entity;
 
-import app.tamingo.BaseEntity;
+import app.tamingo.common.entity.BaseEntity;
 import app.tamingo.domain.schedule.entity.Schedule;
-import app.tamingo.domain.schedule.entity.ScheduleCategory;
 import app.tamingo.domain.todo.enums.RepeatType;
 import app.tamingo.domain.user.entity.User;
 import jakarta.persistence.*;
