+++ src/main/java/app/tamingo/domain/user/entity/User.java
@@ -1,6 +1,6 @@
 package app.tamingo.domain.user.entity;
 
-import app.tamingo.BaseEntity;
+import app.tamingo.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
